perf(portfolio): isolate record workspaces by route
Refs #481
