--- src/apps/portfolio/components/portfolio-record-screen-client.tsx
@@ -1,208 +1,0 @@
-"use client";
-
-import { useMemo, useState } from "react";
-
-import {
-  DegradedStatePanel,
-  MainWithSideRailLayout,
-  WorkbenchPageFrame,
-  WorkbenchSectionStack,
-} from "@/design-system";
-
-import type { PortfolioWorkspace } from "../types";
-import {
-  selectCashflowPartialFailures,
-  selectCashflowWarnings,
-} from "../portfolio-projected-cashflow-view-model";
-import { buildInitialPortfolioControls, buildPortfolioWorkspaceContext } from "../view-model";
-import PortfolioPageLayout from "./portfolio-page-layout";
-import PortfolioPositionsRecordWorkspace from "./portfolio-positions-record-workspace";
-import PortfolioRecordEvidenceRail from "./portfolio-record-evidence-rail";
-import PortfolioScreenRail from "./portfolio-screen-rail";
-import PortfolioHoldingsGrid from "./portfolio-holdings-grid";
-import PortfolioTransactionsRecordWorkspace from "./portfolio-transactions-record-workspace";
-import PortfolioProjectedCashflowModule from "./portfolio-projected-cashflow-module";
-import PortfolioIncomeActivityWorkspace from "./portfolio-income-activity-workspace";
-import PortfolioAllocationPanel from "./portfolio-allocation-panel";
-import type { PortfolioAllocationSelection } from "../types";
-import type { PortfolioRecordScreenKind } from "../portfolio-record-screen-view-model";
-import {
-  buildAllocationHoldingsBreakdown,
-  type AllocationExposureMode,
-} from "../portfolio-allocation-drilldown-view-model";
-import {
-  buildPortfolioRecordDisplayName,
-  buildPortfolioRecordHeaderKpis,
-  buildPortfolioRecordHeaderMeta,
-  buildPortfolioRecordScreenSubtitle,
-  getPortfolioRecordScreenCopy,
-} from "../portfolio-record-screen-view-model";
-
-export default function PortfolioRecordScreenClient({
-  screen,
-  portfolioId,
-  workspace,
-  startDate,
-  endDate,
-}: {
-  screen: PortfolioRecordScreenKind;
-  portfolioId: string | null;
-  workspace: PortfolioWorkspace | null;
-  startDate?: string;
-  endDate?: string;
-}) {
-  const context = useMemo(() => {
-    const controls = buildInitialPortfolioControls(workspace);
-    return buildPortfolioWorkspaceContext(workspace, {
-      ...controls,
-      viewMode: "detailed",
-      columnMode: "expanded",
-    });
-  }, [workspace]);
-  const copy = getPortfolioRecordScreenCopy(screen);
-  const resolvedPortfolioId = portfolioId ?? "No portfolio";
-  const bookDisplayName = workspace ? buildPortfolioRecordDisplayName(workspace) : resolvedPortfolioId;
-  const headerKpis = workspace ? buildPortfolioRecordHeaderKpis(workspace, "30D", screen) : [];
-
-  return (
-    <PortfolioPageLayout>
-      <MainWithSideRailLayout
-        className="portfolio-layout portfolio-record-screen-layout"
-        railClassName="portfolio-screen-rail-shell"
-        mainClassName="portfolio-main portfolio-record-screen-main"
-        rail={<PortfolioScreenRail portfolioId={resolvedPortfolioId} activeScreen={screen} />}
-        side={workspace ? <PortfolioRecordEvidenceRail screen={screen} workspace={workspace} /> : undefined}
-        sideClassName="portfolio-record-evidence-shell"
-        main={
-          <WorkbenchPageFrame
-            className="portfolio-page-frame portfolio-record-page-frame"
-            bodyClassName="portfolio-page-frame-body"
-            title={copy.title}
-            subtitle={buildPortfolioRecordScreenSubtitle(screen)}
-          >
-            <WorkbenchSectionStack className="portfolio-page-sections">
-              {!workspace ? (
-                <DegradedStatePanel
-                  title="Portfolio records unavailable"
-                >
-                  The selected portfolio records are not available for this review.
-                </DegradedStatePanel>
-              ) : (
-                <>
-                  <section className="portfolio-record-standalone-header">
-                    <div>
-                      <span>{copy.kicker}</span>
-                      <h1>{bookDisplayName}</h1>
-                      <p>{buildPortfolioRecordHeaderMeta(workspace)}</p>
-                    </div>
-                    <div className="portfolio-record-standalone-kpis">
-                      {headerKpis.map((kpi) => (
-                        <div key={kpi.label}>
-                          <span>{kpi.label}</span>
-                          <strong>{kpi.value}</strong>
-                        </div>
-                      ))}
-                    </div>
-                  </section>
-                  {screen === "allocation" ? (
-                    <PortfolioAllocationRecordWorkspace
-                      key={workspace.portfolio.portfolio_id}
-                      workspace={workspace}
-                      asOfDate={context.selectedAsOfDate}
-                      reportingCurrency={context.selectedReportingCurrency}
-                    />
-                  ) : null}
-                  {screen === "positions" ? (
-                    <PortfolioPositionsRecordWorkspace
-                      key={workspace.portfolio.portfolio_id}
-                      workspace={workspace}
-                      asOfDate={context.selectedAsOfDate}
-                    />
-                  ) : null}
-                  {screen === "transactions" ? (
-                    <PortfolioTransactionsRecordWorkspace
-                      key={workspace.portfolio.portfolio_id}
-                      workspace={workspace}
-                      asOfDate={context.selectedAsOfDate}
-                      defaultStartDate={startDate ?? context.effectivePeriodStartDate}
-                      defaultEndDate={endDate ?? context.effectivePeriodEndDate}
-                    />
-                  ) : null}
-                  {screen === "income" ? (
-                    <PortfolioIncomeActivityWorkspace workspace={workspace} />
-                  ) : null}
-                  {screen === "cashflow" ? (
-                    <PortfolioProjectedCashflowModule
-                      key={`${workspace.portfolio.portfolio_id}-${context.selectedAsOfDate}`}
-                      portfolioId={workspace.portfolio.portfolio_id}
-                      baseCurrency={workspace.portfolio.base_currency}
-                      asOfDate={context.selectedAsOfDate}
-                      initialCashflowOutlook={workspace.cashflow_outlook}
-                      initialWarnings={selectCashflowWarnings(workspace.warnings)}
-                      initialPartialFailures={selectCashflowPartialFailures(
-                        workspace.partial_failures
-                      )}
-                      defaultExpanded
-                    />
-                  ) : null}
-                </>
-              )}
-            </WorkbenchSectionStack>
-          </WorkbenchPageFrame>
-        }
-      />
-    </PortfolioPageLayout>
-  );
-}
-
-function PortfolioAllocationRecordWorkspace({
-  workspace,
-  asOfDate,
-  reportingCurrency,
-}: {
-  workspace: PortfolioWorkspace;
-  asOfDate: string;
-  reportingCurrency: string;
-}) {
-  const [selectedAllocation, setSelectedAllocation] =
-    useState<PortfolioAllocationSelection | null>(null);
-  const [allocationExposureMode, setAllocationExposureMode] =
-    useState<AllocationExposureMode>("direct");
-  const allocationHoldings = useMemo(
-    () =>
-      buildAllocationHoldingsBreakdown({
-        positions: workspace.positions,
-        cashBalances: workspace.cash_balances ?? [],
-        selection: selectedAllocation,
-        exposureMode: allocationExposureMode,
-      }),
-    [allocationExposureMode, selectedAllocation, workspace.cash_balances, workspace.positions],
-  );
-
-  return (
-    <>
-      <PortfolioAllocationPanel
-        portfolioId={workspace.portfolio.portfolio_id}
-        allocationViews={workspace.allocation_views ?? []}
-        baseCurrency={workspace.portfolio.base_currency}
-        asOfDate={asOfDate}
-        reportingCurrency={reportingCurrency}
-        selectedAllocation={selectedAllocation}
-        onSelectionChange={setSelectedAllocation}
-        onExposureModeChange={setAllocationExposureMode}
-      />
-      <PortfolioHoldingsGrid
-        portfolioId={workspace.portfolio.portfolio_id}
-        positions={allocationHoldings.positions}
-        baseCurrency={workspace.portfolio.base_currency}
-        asOfDate={asOfDate}
-        columnMode="expanded"
-        kicker="Exposure contributors"
-        title={allocationHoldings.title}
-        description={allocationHoldings.description}
-        filterLabel={allocationHoldings.filterLabel}
-        onClearFilter={() => setSelectedAllocation(null)}
-      />
-    </>
-  );
-}
